+++ mcu/attiny-hal/src/attiny167.rs
@@ -59,3 +59,14 @@ pub mod eeprom {
         },
     }
 }
+
+pub mod port {
+    pub use crate::periphals::port::*;
+
+    avr_hal_generic::impl_port_traditional! {
+        enum Ports {
+            A: crate::pac::PORTA = [0, 1, 2, 3, 4, 5, 6, 7],
+            B: crate::pac::PORTB = [0, 1, 2, 3, 4, 5, 6, 7],
+        }
+    }
+}
