+++ mcu/attiny-hal/src/attiny2313.rs
@@ -11,3 +11,15 @@ pub mod eeprom {
         },
     }
 }
+
+pub mod port {
+    pub use crate::periphals::port::*;
+
+    avr_hal_generic::impl_port_traditional! {
+        enum Ports {
+            A: crate::pac::PORTA = [0, 1, 2],
+            B: crate::pac::PORTB = [0, 1, 2, 3, 4, 5, 6, 7],
+            D: crate::pac::PORTD = [0, 1, 2, 3, 4, 5, 6],
+        }
+    }
+}
