+++ mcu/attiny-hal/src/attiny84.rs
@@ -9,3 +9,14 @@ pub mod eeprom {
 
     // Fixme: Implement EEPROM for ATtiny84.
 }
+
+pub mod port {
+    pub use crate::periphals::port::*;
+
+    avr_hal_generic::impl_port_traditional! {
+        enum Ports {
+            A: crate::pac::PORTA = [0, 1, 2, 3, 4, 5, 6, 7],
+            B: crate::pac::PORTB = [0, 1, 2, 3],
+        }
+    }
+}
