+++ mcu/attiny-hal/src/attiny88.rs
@@ -49,3 +49,16 @@ pub mod eeprom {
         },
     }
 }
+
+pub mod port {
+    pub use crate::periphals::port::*;
+
+    avr_hal_generic::impl_port_traditional! {
+        enum Ports {
+            A: crate::pac::PORTA = [0, 1, 2, 3],
+            B: crate::pac::PORTB = [0, 1, 2, 3, 4, 5, 6, 7],
+            C: crate::pac::PORTC = [0, 1, 2, 3, 4, 5, 6, 7],
+            D: crate::pac::PORTD = [0, 1, 2, 3, 4, 5, 6, 7],
+        }
+    }
+}
